requirements.txt
# pyttsx3 is for SystemEngine
pyttsx3==2.98

# azure-cognitiveservices-speech is for AzureEngine
azure-cognitiveservices-speech==1.40.0

# elevenlabs is for ElevenlabsEngine
elevenlabs==1.9.0

# openai is for OpenAIEngine
openai==1.51.2

# gtts is for GTTSEngine
gtts==2.5.3

# coqui_tts is for CoquiEngine
coqui_tts==0.24.2

# stream2sentence is to quickly convert streamed text into sentences for real-time synthesis
stream2sentence==0.2.5

# pydub is used to convert chunks from mp3 to pcm (for openai tts)
pydub==0.25.1

# pyaudio is for playing chunks over output device
pyaudio==0.2.14
PyAudio==0.2.14
faster-whisper==1.0.3
pvporcupine==1.9.5
webrtcvad==2.0.10
halo==0.0.31
torch==2.3.1
torchaudio==2.3.1
scipy==1.12.0
websockets==v12.0
openwakeword==0.6.0
realtimetts[all]
speechrecognition
google-generativeai
fastapi
uvicorn
PyPDF2